showroom/requirements.txt
django>=3.2
django-bakery>=0.13.4
django-crispy-forms>=1.11.2
